Move serializer to seperate file
