+++ conventions-vclib/src/main/resources/vcLibVersions.properties
@@ -3,5 +3,5 @@ uuid=0.8.1
 jsonpath=2.2.0
 jvm.json=20230618
 jvm.cbor=1.18
-eupid=2.1.3-SNAPSHOT
-mdl=1.0.2-SNAPSHOT
+eupid=2.1.3
+mdl=1.0.3
